requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# API Testing
requests==2.31.0

# Mobile Testing (Appium)
Appium-Python-Client==3.1.0
selenium==4.15.2

# Utilities
python-dotenv==1.0.0
allure-pytest==2.13.2

# Reporting
pytest-cov==4.1.0

# Additional utilities
urllib3==2.0.7
certifi==2023.7.22